update-pdfjs.sh: new tool for updating the pdf.js files
As I recently discovered, we can't use the Git submodule system to track the
pdf.js development, so here's a script to make it easy to update to newer
versions. As a comment in the script indicates, it's easy to determine the Git
commit that the downloaded files are associated with.